@@ -1,11 +1,8 @@
-import { FormsModule } from '@angular/forms';
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { HttpClientModule } from '@angular/common/http'
+import { FormsModule } from '@angular/forms';
+import { HttpClientModule } from '@angular/common/http';
 
-
-
-//components
 import { AppComponent } from './app.component';
 import { BuscarImagenComponent } from './components/buscar-imagen/buscar-imagen.component';
 import { ListarImagenComponent } from './components/listar-imagen/listar-imagen.component';
