@@ -1,15 +1,29 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
+import { ImagenService } from 'src/app/services/imagen.service';
 
 @Component({
   selector: 'app-buscar-imagen',
   templateUrl: './buscar-imagen.component.html',
   styleUrls: ['./buscar-imagen.component.css']
 })
-export class BuscarImagenComponent implements OnInit {
+export class BuscarImagenComponent {
+nombreImagen:string;
 
-  constructor() { }
 
-  ngOnInit(): void {
+constructor(private _imagenService: ImagenService) {
+  this.nombreImagen = '';
+}
+
+
+ngOnit():void{
+}
+
+buscarImagenes(){
+  if(this.nombreImagen === ''){
+    this._imagenService.setError('Agrega un texto de busqueda');
+    return;
   }
+  this._imagenService.enviarTerminoBusqueda(this.nombreImagen);
+}
 
 }
